moved components to their own folders along with styling
